Moves version into src
